+++ DataFormats/TrackerRecHit2D/interface/BaseTrackerRecHit.h
@@ -1,7 +1,7 @@
 #ifndef BaseTrackerRecHit_H
 #define BaseTrackerRecHit_H
 
-#include "DataFormats/TrackingRecHit/interface/TrackingRecHit.h"
+#include "DataFormats/TrackerRecHit2D/interface/trackerHitRTTI.h"
 #include "DataFormats/TrackingRecHit/interface/TrackingRecHitGlobalState.h"
 #include "DataFormats/GeometryCommonDetAlgo/interface/ErrorFrameTransformer.h"
 #include "Geometry/CommonDetUnit/interface/TrackerGeomDet.h"
@@ -12,23 +12,6 @@
 //#define VI_DEBUG
 
 class OmniClusterRef;
-
-namespace trackerHitRTTI {
-  // tracking hit can be : single (si1D, si2D, pix), projected, matched or multi
-  enum RTTI { undef=0, single=1, projStereo=2, projMono=3, match=4, multi=5,
-	      fastSingle=6, fastProjStereo=7,fastProjMono=8,fastMatch=9};
-  inline RTTI rtti(TrackingRecHit const & hit)  { return RTTI(hit.getRTTI());}
-  inline bool isUndef(TrackingRecHit const & hit) { return rtti(hit)==undef;}
-  inline bool isSingle(TrackingRecHit const & hit)  { return rtti(hit)==single || rtti(hit)==fastSingle;}
-  inline bool isProjMono(TrackingRecHit const & hit)  { return rtti(hit)==projMono || rtti(hit)==fastProjMono;}
-  inline bool isProjStereo(TrackingRecHit const & hit)  { return rtti(hit)==projStereo || rtti(hit)==fastProjStereo;}
-  inline bool isProjected(TrackingRecHit const & hit)  { return ((rtti(hit)==projMono) | (rtti(hit)==projStereo)) || (rtti(hit)==fastProjMono) | (rtti(hit)==fastProjStereo);}
-  inline bool isMatched(TrackingRecHit const & hit)  { return rtti(hit)==match || rtti(hit)==fastMatch;}
-  inline bool isMulti(TrackingRecHit const & hit)  { return rtti(hit)==multi;}
-  inline bool isSingleType(TrackingRecHit const & hit)  { return (rtti(hit)>0) & (rtti(hit)<4) ;}
-  inline bool isFast(TrackingRecHit const & hit)  { return (rtti(hit)>5) & (rtti(hit)<=9) ;}
-  inline unsigned int  projId(TrackingRecHit const & hit) { return hit.rawId()+int(rtti(hit))-1;}
-}
 
 class BaseTrackerRecHit : public TrackingRecHit { 
 public:
