+++ DataFormats/TrackerRecHit2D/src/SiStripMatchedRecHit2D.cc
@@ -16,8 +16,8 @@ SiStripMatchedRecHit2D::sharesInput( const TrackingRecHit* other,
   }
    
   if (what==all)  return false;
-  // what about multi  etc ???
-  if (trackerHitRTTI::isUndef(*other) ) return false;
+  // what about multi ???
+  if (!trackerHitRTTI::isFromDet(*other) ) return false;
 
   auto const & otherClus = reinterpret_cast<const BaseTrackerRecHit *>(other)->firstClusterRef();
   return (otherClus==stereoClusterRef())  ||  (otherClus==monoClusterRef());
