+++ RecoTracker/DeDx/plugins/DeDxHitInfoProducer.cc
@@ -102,7 +102,7 @@ void DeDxHitInfoProducer::produce(edm::Event& iEvent, const edm::EventSetup& iSe
      auto hb = track.recHitsBegin();
         for(unsigned int h=0;h<track.recHitsSize();h++){
            auto recHit = *(hb+h);
-           if (!recHit->isValid() or  trackerHitRTTI::isUndef(*recHit) ) continue;
+           if (!trackerHitRTTI::isFromDet(*recHit) ) continue;
 
            auto trackDirection = trajParams[h].direction();
            float cosine = trackDirection.z()/trackDirection.mag();
