+++ RecoTracker/TransientTrackingRecHit/interface/Traj2TrackHits.h
@@ -79,7 +79,7 @@ private:
     for(;itm!=e;++itm) {
       auto const & hit = *(*itm).recHit()->hit();
        if( (removeNoDet) & ((*itm).recHitR().det()==nullptr)) continue; 
-      if(trackerHitRTTI::isUndef(hit) | ( hit.dimension()!=2) ) {
+      if(trackerHitRTTI::isUndef(hit) | trackerHitRTTI::isNotFromCluster(hit) | ( hit.dimension()!=2) ) {
 	hits.push_back(hit.clone());
 	continue;
       }
